fourth commit - server validation realized

+++ src/main/java/ru/javamentor/springbootcrud/controller/UsersController.java
@@ -3,12 +3,16 @@ package ru.javamentor.springbootcrud.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import ru.javamentor.springbootcrud.model.User;
 import ru.javamentor.springbootcrud.service.UserService;
+
+import javax.validation.Valid;
+import java.util.List;
 
 
 @Controller
@@ -33,7 +37,11 @@ public class UsersController {
     }
 
     @PostMapping("/create")
-    public String createUser(@ModelAttribute("user") User user) {
+    public String createUser(@ModelAttribute("user") @Valid User user,
+                             BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "userToCreateDetailsForm";
+        }
         userService.addUser(user);
         return "redirect:/users";
     }
@@ -45,16 +53,20 @@ public class UsersController {
 
     @PostMapping("/user")
     public String getUserById(@RequestParam("userId") Long userId, Model model) {
-        if (userId != null) {
-            User user = userService.getUserById(userId);
-            if (user != null) {
-                model.addAttribute("user", user);
-            } else {
-                model.addAttribute("error", "User not found");
-            }
-        } else {
+        if (userId == null) {
             model.addAttribute("error", "User ID parameter is missing");
+            return "userToReadIdForm";
         }
+        if (userId <= 0) {
+            model.addAttribute("error", "User ID should be 1 or more");
+            return "userToReadIdForm";
+        }
+        User user = userService.getUserById(userId);
+        if (user == null) {
+            model.addAttribute("error", "User not found");
+            return "userToReadIdForm";
+        }
+        model.addAttribute("user", user);
         return "userDetails";
     }
 
@@ -66,7 +78,20 @@ public class UsersController {
     @PostMapping("/users")
     public String readUsersList(@RequestParam(value = "count", required = false, defaultValue = "0") int count,
                                 Model model) {
-        model.addAttribute("users", userService.getUsersList(count));
+        if (count == 0) {
+            model.addAttribute("error", "Users List Limit is missing");
+            return "usersListLimitForm";
+        }
+        if (count < 0) {
+            model.addAttribute("error", "User List Limit should be 1 or more");
+            return "usersListLimitForm";
+        }
+        List<User> users = userService.getUsersList(count);
+        if (users.isEmpty()) {
+            model.addAttribute("error", "Users List not found");
+            return "usersListLimitForm";
+        }
+        model.addAttribute("users", users);
         return "usersList";
     }
 
@@ -77,21 +102,29 @@ public class UsersController {
 
     @PostMapping("/editedUser")
     public String getEditedUserById(@RequestParam("userId") Long userId, Model model) {
-        if (userId != null) {
-            User user = userService.getUserById(userId);
-            if (user != null) {
-                model.addAttribute("user", user);
-            } else {
-                model.addAttribute("error", "User not found");
-            }
-        } else {
+        if (userId == null) {
             model.addAttribute("error", "User ID parameter is missing");
+            return "userToUpdateDetailsForm";
         }
+        if (userId <= 0) {
+            model.addAttribute("error", "User ID should be 1 or more");
+            return "userToUpdateDetailsForm";
+        }
+        User user = userService.getUserById(userId);
+        if (user == null) {
+            model.addAttribute("error", "User not found");
+            return "userToUpdateDetailsForm";
+        }
+        model.addAttribute("user", user);
         return "userToUpdateDetailsForm";
     }
 
     @PostMapping("/update")
-    public String editUser(@ModelAttribute("user") User user) {
+    public String editUser(@ModelAttribute("user") @Valid User user,
+                           BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return "userToUpdateDetailsForm";
+        }
         userService.updateUser(user);
         return "redirect:/users";
     }
@@ -103,7 +136,20 @@ public class UsersController {
 
 
     @PostMapping("/deleteUser")
-    public String deleteUser(@ModelAttribute("userId") Long userId) {
+    public String deleteUser(@RequestParam("userId") Long userId, Model model) {
+        if (userId == null) {
+            model.addAttribute("error", "User ID parameter is missing");
+            return "userToDeleteIdForm";
+        }
+        if (userId <= 0) {
+            model.addAttribute("error", "User ID should be 1 or more");
+            return "userToDeleteIdForm";
+        }
+        User user = userService.getUserById(userId);
+        if (user == null) {
+            model.addAttribute("error", "User not found");
+            return "userToDeleteIdForm";
+        }
         userService.deleteUserById(userId);
         return "redirect:/users";
     }
